Add `alpha.h` to hold elemental `zen` utilities
At the moment make_kaizen.py creates kaizen.h by dumping the contents of the constituent header files sorted in alphabetical order, so the name 'alpha.h' of the file is chosen such that its contents appear first in kaizen.h, so any downstream code can use it. This slighly breaks the principle of independence of all individual headers, but only to a first degree and seems to bring benefits that outweight the cost of a slightly increased coupling.