Bifurcate `quote()` into `squote()` and `dquote()`
Introduces two new inline functions, `dquote` and `squote`, for quoting strings in double and single quotes, respectively. The existing `quote` function is redefined as a macro, which can be set to either based on user-preference. NOTE: existing Kaizen tests currently assume `quote` to be `dquote` (probably worth changing this assumption to be `squote`.